requirements.txt
flask
flask-cors
Pillow
numpy
scikit-learn
joblib
pandas